refact: move wait task to the client level
